Removed unused functions from main.jsx
